pyproject.toml
[tool.poetry]
name = "cbp-one"
version = "0.1.0"
description = "CBP One Mexico Telegram Bot"
authors = ["Ian <exor2008@yandex.ru>"]
readme = "README.md"
packages = [{include = "cbp_one"}]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^20.1"
dataclasses-json = "^0.5.7"


[tool.poetry.group.dev.dependencies]
mypy = "^1.1.1"
isort = "^5.12.0"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 100
profile = "black"